# specs/001-perf-problem-simulator/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Performance Problem Simulator API
  description: |
    API for triggering and monitoring intentional performance problems in an Azure App Service
    application for educational and diagnostic demonstration purposes.
    
    ⚠️ **WARNING**: This API intentionally creates performance problems. It is designed for
    testing and learning purposes only. Do not use in production environments.
    
    ## Endpoint Categories
    
    - **Problem Triggers**: Endpoints that start performance problems
    - **Metrics**: Endpoints that report current application state
    - **Control**: Endpoints for managing simulations
    
    ## API Style
    
    This API uses action-oriented RPC-style endpoints (e.g., `/api/trigger-high-cpu`) rather than
    REST resource-based endpoints, as the operations are actions rather than resource management.
  version: 1.0.0
  contact:
    name: Performance Problem Simulator
  license:
    name: MIT
servers:
  - url: https://localhost:5001
    description: Local development
  - url: https://{appname}.azurewebsites.net
    description: Azure App Service
    variables:
      appname:
        default: perfsimasp48
        description: Azure App Service name
...
paths:
...
  /api/loadtest:
    get:
      tags:
        - LoadTest
      summary: Execute load test request
      description: |
        Endpoint designed for Azure Load Testing. Simulates realistic application
        behavior that degrades gracefully under concurrent load.
        
        **Behavior:**
        - Performs sustained CPU work (SHA256 hash iterations)
        - Holds memory buffer for entire request duration
        - Applies degradation delay when concurrent requests exceed soft limit
        - Throws random exceptions after 120 seconds (20% probability)
        
        **Use Cases:**
        - Azure Load Testing URL-based tests (no JMeter script required)
        - Performance baseline establishment
        - Thread pool exhaustion testing
      operationId: executeLoadTest
      parameters:
        - name: workIterations
          in: query
          description: SHA256 iterations per CPU work cycle. Controls sustained CPU usage.
          schema:
            type: integer
            default: 200
            minimum: 0
        - name: bufferSizeKb
          in: query
          description: Memory buffer held for entire request duration (KB).
          schema:
            type: integer
            default: 20000
            minimum: 0
        - name: baselineDelayMs
          in: query
          description: Minimum request duration in milliseconds.
          schema:
            type: integer
            default: 500
            minimum: 0
        - name: softLimit
          in: query
          description: Concurrent requests before degradation begins.
          schema:
            type: integer
            default: 25
            minimum: 1
        - name: degradationFactor
          in: query
          description: Delay milliseconds added per request over soft limit.
          schema:
            type: integer
            default: 500
            minimum: 0
      responses:
        '200':
          description: Load test request completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadTestResult'
        '500':
          description: Random exception thrown (expected after 120s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadTestResult'
...
components:
  schemas:
...
    LoadTestResult:
      type: object
      required:
        - elapsedMs
        - concurrentRequestsAtStart
        - degradationDelayAppliedMs
        - workIterationsCompleted
        - memoryAllocatedBytes
        - workCompleted
        - exceptionThrown
        - timestamp
      properties:
        elapsedMs:
          type: integer
          format: int64
          description: Total request duration in milliseconds
        concurrentRequestsAtStart:
          type: integer
          description: Number of concurrent requests when this request started
        degradationDelayAppliedMs:
          type: integer
          description: Additional delay applied due to exceeding soft limit
        workIterationsCompleted:
          type: integer
          description: SHA256 iterations completed
        memoryAllocatedBytes:
          type: integer
          format: int64
          description: Memory buffer size allocated
        workCompleted:
          type: boolean
          description: Whether all work completed successfully
        exceptionThrown:
          type: boolean
          description: Whether a random exception was thrown
        exceptionType:
          type: string
          nullable: true
          description: Exception type if thrown
        timestamp:
          type: string
          format: date-time
...